Split network data sources into separate files
